backend/requirements.txt
email-validator==1.1.2
emails==0.6
fastapi==0.63.0
gunicorn==20.1.0
pandas==1.2.3
passlib[bcrypt]==1.7.4
python-jose==3.2.0
python-multipart==0.0.5
SQLAlchemy==1.4.3
uvicorn==0.13.4
